Shorter titles.
If I am correct, this variable controls the page title amended to each guide after it's actual title. As we learn more about how to get our pages to rank better, we've discovered that having long titles, past a certain character limit, is a detriment. I think that the title of a guide coupled with the word Linode is enough to tell the reader where they are, and this gives us more breathing room to make titles that accurately describe the guide.