Prune libraries and tools
This commit removes all libraries from the site that fit the following
conditions:

* Have not received a commit within the last year (and are likely not working)
* Are targeting the API v3

The second point might be contentious, but I think we should stop
recommending people use the v3 API, as it will only make it harder to
move away from it.

Please close this PR if I'm being too zealous.